Fix|Homepage Clean up
