+++ frontend/src/components/Layout.jsx
@@ -1,76 +1,8 @@
 import React from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { Button } from '@/components/ui/button';
-import { useAuth } from '@/contexts/AuthContext';
-import { Heart, MessageSquare, FileText, LogOut, Stethoscope } from 'lucide-react';
-import logoImage from '@/assets/logo.png';
 
 export const Layout = ({ children }) => {
-  const { user, logout } = useAuth();
-  const location = useLocation();
-
-  const navigation = [
-    { name: 'Dashboard', href: '/', icon: Heart },
-    { name: 'Risk Assessment', href: '/assessment', icon: Stethoscope },
-    { name: 'Chat Assistant', href: '/chat', icon: MessageSquare },
-    { name: 'FAQ', href: '/faq', icon: FileText },
-  ];
-
   return (
     <div className="min-h-screen bg-background">
-      <header className="bg-card shadow-card border-b border-border">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center h-16">
-            <div className="flex items-center space-x-8">
-              <Link to="/" className="flex items-center space-x-3">
-                <img src={logoImage} alt="DiabetesPredict" className="w-8 h-8" />
-                <span className="text-xl font-bold text-foreground">DiabetesPredict</span>
-              </Link>
-              
-              <nav className="hidden md:flex space-x-6">
-                {navigation.map((item) => {
-                  const Icon = item.icon;
-                  const isActive = location.pathname === item.href;
-                  return (
-                    <Link
-                      key={item.name}
-                      to={item.href}
-                      className={`flex items-center space-x-1 px-3 py-2 rounded-lg text-sm font-medium transition-colors ${
-                        isActive
-                          ? 'text-primary bg-secondary'
-                          : 'text-muted-foreground hover:text-foreground hover:bg-secondary'
-                      }`}
-                    >
-                      <Icon className="h-4 w-4" />
-                      <span>{item.name}</span>
-                    </Link>
-                  );
-                })}
-              </nav>
-            </div>
-
-            <div className="flex items-center space-x-4">
-              {user && (
-                <>
-                  <span className="text-sm text-muted-foreground">
-                    Welcome, {user.name}
-                  </span>
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={logout}
-                    className="flex items-center space-x-1"
-                  >
-                    <LogOut className="h-4 w-4" />
-                    <span>Logout</span>
-                  </Button>
-                </>
-              )}
-            </div>
-          </div>
-        </div>
-      </header>
-
       <main className="flex-1">
         {children}
       </main>
