+++ frontend/src/contexts/AuthConstants.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const AuthContext = createContext(undefined);
+
+export default AuthContext;
